+++ src/rules/validation/deployment/RuleBlacklistOwnable2Step.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.8.20;
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {Ownable2Step} from "@openzeppelin/contracts/access/Ownable2Step.sol";
 import {Context} from "@openzeppelin/contracts/utils/Context.sol";
+import {Ownable2StepERC165Module} from "../../../modules/Ownable2StepERC165Module.sol";
 import {RuleBlacklistBase} from "../abstract/base/RuleBlacklistBase.sol";
 import {RuleAddressSet} from "../abstract/RuleAddressSet/RuleAddressSet.sol";
 
@@ -11,7 +12,7 @@ import {RuleAddressSet} from "../abstract/RuleAddressSet/RuleAddressSet.sol";
  * @title RuleBlacklistOwnable2Step
  * @notice Ownable2Step variant of RuleBlacklist with owner-based authorization hooks.
  */
-contract RuleBlacklistOwnable2Step is RuleBlacklistBase, Ownable2Step {
+contract RuleBlacklistOwnable2Step is RuleBlacklistBase, Ownable2Step, Ownable2StepERC165Module {
     /*//////////////////////////////////////////////////////////////
                              CONSTRUCTOR
     //////////////////////////////////////////////////////////////*/
@@ -25,6 +26,17 @@ contract RuleBlacklistOwnable2Step is RuleBlacklistBase, Ownable2Step {
     function _authorizeAddressListAdd() internal view virtual override onlyOwner {}
 
     function _authorizeAddressListRemove() internal view virtual override onlyOwner {}
+
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(RuleBlacklistBase, Ownable2StepERC165Module)
+        returns (bool)
+    {
+        return Ownable2StepERC165Module.supportsInterface(interfaceId)
+            || RuleBlacklistBase.supportsInterface(interfaceId);
+    }
 
     /*//////////////////////////////////////////////////////////////
                         INTERNAL FUNCTIONS