+++ src/rules/validation/deployment/RuleERC2980Ownable2Step.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.8.20;
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {Ownable2Step} from "@openzeppelin/contracts/access/Ownable2Step.sol";
 import {Context} from "@openzeppelin/contracts/utils/Context.sol";
+import {Ownable2StepERC165Module} from "../../../modules/Ownable2StepERC165Module.sol";
 import {RuleERC2980Base} from "../abstract/base/RuleERC2980Base.sol";
 
 /**
@@ -11,7 +12,7 @@ import {RuleERC2980Base} from "../abstract/base/RuleERC2980Base.sol";
  * @notice Ownable2Step variant of RuleERC2980 with owner-based authorization hooks.
  * @dev All whitelist and frozenlist management functions are restricted to the contract owner.
  */
-contract RuleERC2980Ownable2Step is RuleERC2980Base, Ownable2Step {
+contract RuleERC2980Ownable2Step is RuleERC2980Base, Ownable2Step, Ownable2StepERC165Module {
     /*//////////////////////////////////////////////////////////////
                              CONSTRUCTOR
     //////////////////////////////////////////////////////////////*/
@@ -37,6 +38,17 @@ contract RuleERC2980Ownable2Step is RuleERC2980Base, Ownable2Step {
     function _authorizeFrozenlistAdd() internal view virtual override onlyOwner {}
 
     function _authorizeFrozenlistRemove() internal view virtual override onlyOwner {}
+
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(RuleERC2980Base, Ownable2StepERC165Module)
+        returns (bool)
+    {
+        return Ownable2StepERC165Module.supportsInterface(interfaceId)
+            || RuleERC2980Base.supportsInterface(interfaceId);
+    }
 
     /*//////////////////////////////////////////////////////////////
                         INTERNAL FUNCTIONS