+++ src/rules/validation/deployment/RuleIdentityRegistryOwnable2Step.sol
@@ -3,13 +3,15 @@ pragma solidity ^0.8.20;
 
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {Ownable2Step} from "@openzeppelin/contracts/access/Ownable2Step.sol";
+import {Ownable2StepERC165Module} from "../../../modules/Ownable2StepERC165Module.sol";
+import {RuleTransferValidation} from "../abstract/core/RuleTransferValidation.sol";
 import {RuleIdentityRegistryBase} from "../abstract/base/RuleIdentityRegistryBase.sol";
 
 /**
  * @title RuleIdentityRegistryOwnable2Step
  * @notice Ownable2Step variant of RuleIdentityRegistry.
  */
-contract RuleIdentityRegistryOwnable2Step is RuleIdentityRegistryBase, Ownable2Step {
+contract RuleIdentityRegistryOwnable2Step is RuleIdentityRegistryBase, Ownable2Step, Ownable2StepERC165Module {
     /*//////////////////////////////////////////////////////////////
                              CONSTRUCTOR
     //////////////////////////////////////////////////////////////*/
@@ -21,4 +23,15 @@ contract RuleIdentityRegistryOwnable2Step is RuleIdentityRegistryBase, Ownable2S
     //////////////////////////////////////////////////////////////*/
 
     function _authorizeIdentityRegistryManager() internal view virtual override onlyOwner {}
+
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(RuleTransferValidation, Ownable2StepERC165Module)
+        returns (bool)
+    {
+        return Ownable2StepERC165Module.supportsInterface(interfaceId)
+            || RuleTransferValidation.supportsInterface(interfaceId);
+    }
 }