+++ src/rules/validation/deployment/RuleSanctionsListOwnable2Step.sol
@@ -4,7 +4,9 @@ pragma solidity ^0.8.20;
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {Ownable2Step} from "@openzeppelin/contracts/access/Ownable2Step.sol";
 import {Context} from "@openzeppelin/contracts/utils/Context.sol";
+import {Ownable2StepERC165Module} from "../../../modules/Ownable2StepERC165Module.sol";
 import {ERC2771Context} from "../../../modules/MetaTxModuleStandalone.sol";
+import {RuleTransferValidation} from "../abstract/core/RuleTransferValidation.sol";
 import {RuleSanctionsListBase} from "../abstract/base/RuleSanctionsListBase.sol";
 import {ISanctionsList} from "../../interfaces/ISanctionsList.sol";
 
@@ -12,7 +14,7 @@ import {ISanctionsList} from "../../interfaces/ISanctionsList.sol";
  * @title RuleSanctionsListOwnable2Step
  * @notice Ownable2Step variant of RuleSanctionsList.
  */
-contract RuleSanctionsListOwnable2Step is RuleSanctionsListBase, Ownable2Step {
+contract RuleSanctionsListOwnable2Step is RuleSanctionsListBase, Ownable2Step, Ownable2StepERC165Module {
     /*//////////////////////////////////////////////////////////////
                              CONSTRUCTOR
     //////////////////////////////////////////////////////////////*/
@@ -27,6 +29,17 @@ contract RuleSanctionsListOwnable2Step is RuleSanctionsListBase, Ownable2Step {
     //////////////////////////////////////////////////////////////*/
 
     function _authorizeSanctionListManager() internal view virtual override onlyOwner {}
+
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(RuleTransferValidation, Ownable2StepERC165Module)
+        returns (bool)
+    {
+        return Ownable2StepERC165Module.supportsInterface(interfaceId)
+            || RuleTransferValidation.supportsInterface(interfaceId);
+    }
 
     /*//////////////////////////////////////////////////////////////
                         INTERNAL FUNCTIONS