+++ src/rules/validation/deployment/RuleWhitelistOwnable2Step.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.8.20;
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 import {Ownable2Step} from "@openzeppelin/contracts/access/Ownable2Step.sol";
 import {Context} from "@openzeppelin/contracts/utils/Context.sol";
+import {Ownable2StepERC165Module} from "../../../modules/Ownable2StepERC165Module.sol";
 import {RuleWhitelistBase} from "../abstract/base/RuleWhitelistBase.sol";
 import {RuleAddressSet} from "../abstract/RuleAddressSet/RuleAddressSet.sol";
 
@@ -11,7 +12,7 @@ import {RuleAddressSet} from "../abstract/RuleAddressSet/RuleAddressSet.sol";
  * @title RuleWhitelistOwnable2Step
  * @notice Ownable2Step variant of RuleWhitelist with owner-based authorization hooks.
  */
-contract RuleWhitelistOwnable2Step is RuleWhitelistBase, Ownable2Step {
+contract RuleWhitelistOwnable2Step is RuleWhitelistBase, Ownable2Step, Ownable2StepERC165Module {
     /*//////////////////////////////////////////////////////////////
                              CONSTRUCTOR
     //////////////////////////////////////////////////////////////*/
@@ -36,6 +37,17 @@ contract RuleWhitelistOwnable2Step is RuleWhitelistBase, Ownable2Step {
     function _authorizeAddressListRemove() internal view virtual override onlyOwner {}
 
     function _authorizeCheckSpenderManager() internal view virtual override onlyOwner {}
+
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(RuleWhitelistBase, Ownable2StepERC165Module)
+        returns (bool)
+    {
+        return Ownable2StepERC165Module.supportsInterface(interfaceId)
+            || RuleWhitelistBase.supportsInterface(interfaceId);
+    }
 
     /*//////////////////////////////////////////////////////////////
                         INTERNAL FUNCTIONS